convert float to int

+++ jwst/skymatch/skyimage.py
@@ -446,8 +446,8 @@ None, optional
 
                 # set pixels in 'fill_mask' that are inside a polygon to True:
                 x, y = self.wcs_inv(ra, dec)
-                xcheck = np.ndarray.round(x)
-                ycheck = np.ndarray.round(y)
+                xcheck = np.ndarray.round(x).astype(int)
+                ycheck = np.ndarray.round(y).astype(int)
                 if np.min(xcheck) == np.max(xcheck) or np.min(ycheck) == np.max(ycheck):
                     continue
                 poly_vert = list(zip(*[x, y]))
